Rearrange HITSKleinberg and provide Inspector for HITSKleinbergValue
