+++ arangod/Pregel/Algos/CMakeLists.txt
@@ -1,7 +1,6 @@
 target_sources(arango_pregel PRIVATE
   ConnectedComponents.cpp
   HITS.cpp
-  HITSKleinberg.cpp
   ColorPropagation.cpp)
 
 if (USE_MAINTAINER_MODE)
@@ -10,6 +9,7 @@ endif()
 
 add_subdirectory(DMID)
 add_subdirectory(EffectiveCloseness)
+add_subdirectory(HITSKleinberg)
 add_subdirectory(LabelPropagation)
 add_subdirectory(LineRank)
 add_subdirectory(PageRank)
