+++ arangod/Pregel/Worker/GraphStore.cpp
@@ -54,6 +54,7 @@
 #include "VocBase/ticks.h"
 #include "VocBase/vocbase.h"
 
+#include "Pregel/Algos/HITSKleinberg/HITSKleinbergValue.h"
 #include "Pregel/Algos/LabelPropagation/LPValue.h"
 #include "Pregel/Algos/SCC/SCCValue.h"
 #include "Pregel/Algos/SLPA/SLPAValue.h"
@@ -819,7 +820,7 @@ template class arangodb::pregel::GraphStore<algos::WCCValue, uint64_t>;
 template class arangodb::pregel::GraphStore<algos::SCCValue, int8_t>;
 template class arangodb::pregel::GraphStore<ECValue, int8_t>;
 template class arangodb::pregel::GraphStore<HITSValue, int8_t>;
-template class arangodb::pregel::GraphStore<HITSKleinbergValue, int8_t>;
+template class arangodb::pregel::GraphStore<algos::HITSKleinbergValue, int8_t>;
 template class arangodb::pregel::GraphStore<DMIDValue, float>;
 template class arangodb::pregel::GraphStore<algos::LPValue, int8_t>;
 template class arangodb::pregel::GraphStore<algos::SLPAValue, int8_t>;
